requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
Django==4.0.1
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.2
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
nest-asyncio==1.5.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.25
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
